Revert "Fix implicit signed conversion and sign comparison"
This reverts commit b4e83367d2de00c1c74facd0bcc6592d0a26a6ba.

This introduced regressions in the TLS module as reported in GitHub
issue #271. Reverting until this can be examined in more detail.

+++ src/buffer.h
@@ -42,7 +42,7 @@ struct Buffer {
     size_t rx_bytes;
 };
 
-struct Buffer *new_buffer(size_t, struct ev_loop *);
+struct Buffer *new_buffer(int, struct ev_loop *);
 void free_buffer(struct Buffer *);
 
 ssize_t buffer_recv(struct Buffer *, int, int, struct ev_loop *);
